Tasks
This update adds tasks. They provide isolation to prevent error
cascading and CPU usage control. A room task is created for which
it is credited with CPU. It has a maximum bucket. If it exceeds
this it creates a negative amount which must be recovered before
the task can run again. The persistance happens with the unique
name for the task. Not all tasks require a unique name if they do
not require a credit account. Some tasks can be spawned as isolated
which means they have their own credit account.

This means tasks, rooms, or creeps which happen to go over budget
are required to wait out the next one or more ticks until they
recover what they spent in CPU.