Miners Build Containers
The miners build containers next to their designated sources. The
miner do not actually build then but they create the construction
site for which the general workers build, they locate themselves
onto any existing container- even if it was not built by them as
long as its adjacent to the source.