Decision Tree Now Supports Additional Parameter
A big change to the decision tree now supports an additional
parameter. This is currently called oneshot but as an after
thought I realized it could be overloaded for a lot of purposes.

I actually might go ahead and rework it to be a dictionary of
options whereas one option is oneshot. This says to the general
worker to only pull or push energy/resources once then
re-eveluate the tree again. It prevent, for example, the hauler
from nursing an object until its store is empty.