Rollup & New Creep (StateMachineCreep)
The new creep type `StateMachineCreep` is designed to handle cases
where the `GeneralWorkerCreep` is insufficient. The state machine
allows tasks to be pushed and then executed in order. It also supports
the grouping of state pushes into sets. Either the entire set is pushes
or none of the set and each set has a unique identifier such that only
one of this unique identifier can exist as a set at any given time.