Hauler Code Only Looks At Energy Near Sources
The hauler code doubles the hauling capacity by doubling the number
of haulers when total dropped energy in a room exceeds 200. Now, it
only looks at dropped energy adjacent to sources. This is particularly
in response to the source miner code.

However, I may create a inter-module communcation API. This way the
miner can register that there is a problem and the hauler could
register to hook that instead of just looking for dropped energy.