Fixed Bug for Claim Team Using Allowable Room List
The initial claim team phase is to build a creep with claim parts
and this creep obeyed the allowed rooms list. The second phase creep
helps to build the spawn _BUT_ it did not obey the allowable rooms
causing it to enter into unfriendly territory. Now, it properly
looks at that list and only uses rooms marked as safe IF a list is
provided.