Lots of Refactoring and Fixes in Industry Code
The industry code isn't modular at all. It has bugs. It isn't
flexible. The code doesn't look beautiful because the functions
are way too large and long where I slammed them together just
trying to get things rolling and tested.

Now, this is part of the work to essentially refactor the labrat
module so the code is cleaner and more ergonomic to work with.