Check Mover Creep's Store
Somehow the labman's orders get stuck because there isn't any of
a needed input. So, the system sits there waiting for it. This
bug fixes an issue in the logic to detect if any of the input
exists. It fixes an issue where the mover creep needed to be
checked.