Wall Repair and Controller Link
Towers Repair Walls and Ramparts

Towers now repair the lowest wall/rampart but do so
at a regulated amount. TODO> There needs to be better
control on the flow of energy. A specific value would
be better.

Link From Miners To Controller

+ furthest miner now completes orders from room to
  send energy to link at controller if it exists
+ room now creates energy orders for controller link
+ upgrader now looks for link and pulls energy from it

The rationale behind this change is that we need to reduce
CPU usage. The best way is to send energy directly from
the source to the controller via links. It eliminates the
pathing CPU usage from the hauler. Also, since at RCL 6
there is no worker around without construction sites this
allow a reduction in the work load on the hauler.